fix paths after the rename
