+++ src/api/scope/lib/fetch.ts
@@ -5,7 +5,7 @@ import HooksManager from '../../../hooks';
 // import logger from '../../../logger/logger';
 import ScopeComponentsImporter from '../../../scope/component-ops/scope-components-importer';
 import ComponentObjects from '../../../scope/component-objects';
-import CompsAndLanesObjects from '../../../scope/objects-to-push';
+import CompsAndLanesObjects from '../../../scope/comps-and-lanes-objects';
 import LaneObjects from '../../../scope/lane-objects';
 
 const HooksManagerInstance = HooksManager.getInstance();
