+++ src/consumer/component/component-loader.ts
@@ -11,7 +11,7 @@ import { DependencyResolver, updateDependenciesVersions } from './dependencies/d
 import { getScopeRemotes } from '../../scope/scope-remotes';
 import { ModelComponent } from '../../scope/models';
 import ComponentsPendingImport from '../component-ops/exceptions/components-pending-import';
-import CompsAndLanesObjects from '../../scope/objects-to-push';
+import CompsAndLanesObjects from '../../scope/comps-and-lanes-objects';
 
 export default class ComponentLoader {
   _componentsCache: Record<string, any> = {}; // cache loaded components
